requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
alembic>=1.13.1
psycopg2-binary>=2.9.9
celery>=5.4.0
redis>=5.0.4
playwright>=1.44.0
python-telegram-bot>=21.1.1
pgvector>=0.2.5
pydantic>=2.7.2
pydantic-settings>=2.2.1
groq>=0.9.0
cerebras_cloud_sdk>=1.0.0
pytest>=8.2.1
pytest-asyncio>=0.23.6
pytest-mock>=3.14.0
pytest-cov>=5.0.0
httpx>=0.27.0
python-dotenv>=1.0.1
